Crate pallet_object_store

Pallet object store, used for simple object storage on the network.

Re-exports

pub use __tt_error_token_1 as tt_error_token;

pub use __tt_default_parts_7 as tt_default_parts;

Structs

CallObject

Mapping to the object offset and size within an extrinsic

Pallet

Pallet object-store, used for storing arbitrary user-provided data combined into object-store.

Enums

Call

Contains one variant per dispatchable that can be called by an extrinsic.

Event

pallet-object-store events

Traits

Config

Configuration trait of this pallet.

Type Definitions

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.