Crate pallet_object_store

Pallet object store, used for simple object storage on the network.

Re-exports

pub use __tt_error_token_1 as tt_error_token;

pub use __tt_default_parts_7 as tt_default_parts;

pub use __tt_extra_parts_7 as tt_extra_parts;

pub use __tt_default_parts_v2_7 as tt_default_parts_v2;

Modules

dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Structs

CallObject

Mapping to the object offset and size within an extrinsic

Pallet

Pallet object-store, used for storing arbitrary user-provided data combined into object-store.

Enums

Call

Contains a variant per dispatchable extrinsic that this pallet has.

Event

pallet-object-store events

Traits

Config

Configuration trait of this pallet.

Type Aliases

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.